requirements.txt
Pillow
duckduckgo_search
requests